+++ hc/api/views.py
@@ -92,6 +92,9 @@ def _update(check, spec):
     if "desc" in spec:
         check.desc = spec["desc"]
 
+    if "manual_resume" in spec:
+        check.manual_resume = spec["manual_resume"]
+
     if "timeout" in spec and "schedule" not in spec:
         check.kind = "simple"
         check.timeout = td(seconds=spec["timeout"])