+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Bug Fixes
+- Removing Pager Team integration, project appears to be discontinued
+
 ## v1.15.0 - 2020-06-04
 
 ### Improvements