+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Add "Get a list of checks's logged pings" API call (#371)
 - The /api/v1/checks/ endpoint now accepts either UUID or `unique_key` (#370)
 - Added /api/v1/checks/uuid/flips/ endpoint (#349)
+- In the cron expression dialog, show a human-friendly version of the expression
 
 
 ### Bug Fixes